[BE] REFACTOR: router 분기 처리
