+++ server/package.json
@@ -17,7 +17,8 @@
       "^@models(.*)$": "<rootDir>/src/models/$1",
       "^@models(.*)/$": "<rootDir>/src/models/$1",
       "^@loaders/(.*)$": "<rootDir>/src/loaders/$1",
-      "^@utils/(.*)$": "<rootDir>/src/utils/$1"
+      "^@utils/(.*)$": "<rootDir>/src/utils/$1",
+      "^@routes/(.*)$": "<rootDir>/src/routes/$1"
     }
   },
   "dependencies": {
@@ -50,6 +51,7 @@
     "@config": "./src/config",
     "@models": "./src/models",
     "@loaders": "./src/loaders",
-    "@utils": "./src/utils"
+    "@utils": "./src/utils",
+    "@routes" : "./src/routes"
   }
 }
