+++ server/src/loaders/express.js
@@ -8,7 +8,7 @@ const expressLoader = app => {
   app.use(express.json());
   app.use(express.urlencoded({ extended: false }));
 
-  app.use('/', indexRouter);
+  app.use('/api', indexRouter);
 };
 
 module.exports = expressLoader;
